+++ src/emsi/resource/contact.ts
@@ -7,18 +7,18 @@ export class Contact extends Default {
     contact: ContactType
     detail: Detail
 
-    constructor(contact: ContactType, detail: Detail) {
+    constructor (contact: ContactType, detail: Detail) {
         ResourceError.checkLength(detail, MAX_DETAIL_LENGTH)
         super()
         this.contact = contact
         this.detail = detail 
     }
 
-    static default(): Contact {
+    static default (): Contact {
         return new Contact(ContactType.IP_ADDRESS, '0.0.0.0')
     }
 
-    assign(source: Record<string, any>): this {
+    assign (source: Record<string, any>): this {
         let key
         const keys = Object.keys(source)