+++ src/error/index.ts
@@ -3,3 +3,4 @@ export { EventError } from './event.error'
 export { PositionError } from './position.error'
 export { EmsiError } from './emsi.error'
 export { MissionError } from './mission.error'
+export { ResourceError } from "./resource.error"