+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.1 - 2021-05-26
+
+- Fix EGeo.addWeather error
+- Fix Event.setOccDatime syntax
+- Fix Event.addCasualties syntax
+- Change add functions plural
+- Publish yarn.lock
+
 # 0.1.0 - 2021-05-18
 
 ✨ First release